// Application.API/ClientApp/src/app/models/person.ts
export interface Person{
    firstName: string;
    lastName: string;
    emailAddress: string;
    mobileNumber: string;
    customerId: number;
    businessArea1: number;
    businessArea2: number;
    businessArea3: number;
    dateCreated: Date;
    createdBy: string;
    dateModified: Date;
    modifiedBy: string;

}